Address GLTF cache review feedback
