+++ tests/gltf-model-cache.test.ts
@@ -1,9 +1,5 @@
 import { expect, test } from "bun:test"
-import * as THREE from "three"
-import {
-  cloneSceneWithIndependentMaterials,
-  normalizeGltfModelCacheKey,
-} from "../src/utils/gltf-model-cache"
+import { normalizeGltfModelCacheKey } from "../src/utils/gltf-model-cache"
 
 test("normalizes cachebust_origin without dropping meaningful query params", () => {
   expect(
@@ -15,19 +11,4 @@ test("normalizes cachebust_origin without dropping meaningful query params", () 
   expect(
     normalizeGltfModelCacheKey("/models/chip.glb?cachebust_origin=abc"),
   ).toBe("/models/chip.glb")
-})
-
-test("clones cached scenes with independent material instances", () => {
-  const material = new THREE.MeshStandardMaterial({ color: 0x00ff00 })
-  const original = new THREE.Group()
-  original.add(new THREE.Mesh(new THREE.BoxGeometry(1, 1, 1), material))
-
-  const firstClone = cloneSceneWithIndependentMaterials(original)
-  const secondClone = cloneSceneWithIndependentMaterials(original)
-  const firstMesh = firstClone.children[0] as THREE.Mesh
-  const secondMesh = secondClone.children[0] as THREE.Mesh
-
-  expect(firstMesh.material).not.toBe(material)
-  expect(secondMesh.material).not.toBe(material)
-  expect(firstMesh.material).not.toBe(secondMesh.material)
 })
